+++ CHANGELOG.md
@@ -17,6 +17,7 @@ An EtherCAT master written in Rust.
 
 ### Changed
 
+- **(breaking)** [#TODO] `send_frames_blocking` now gives the raw Ethernet II frame to the send closure instead of a `SendableFrame`.
 - **(breaking)** [#30] Removed `PduError::Encode` variant.
 - **(breaking)** [#25] Changed `pdu_rx` to `receive_frame` to mirror `send_frames_blocking`.
 - **(breaking)** [#20] Changed the way the client, tx and rx instances are initialised to only allow